enforce @inaccessible and @tag names in the subgraph
In order to enforce consistency in the supergraph, `@inaccessible` and `@tag` directives have to be named consistently across subgraphs and cannot be renamed.

Resolves: https://github.com/apollographql/federation-jvm/issues/233
