--- graphql-java-support/src/test/resources/schemas/polymorphicSubgraph_serviceSDL.graphql
@@ -1,58 +1,0 @@
-schema @link(import : ["@key"], url : "https://specs.apollo.dev/federation/v2.0"){
-  query: Query
-}
-
-directive @federation__extends on OBJECT | INTERFACE
-
-directive @federation__external on OBJECT | FIELD_DEFINITION
-
-directive @federation__inaccessible on SCALAR | OBJECT | FIELD_DEFINITION | ARGUMENT_DEFINITION | INTERFACE | UNION | ENUM | ENUM_VALUE | INPUT_OBJECT | INPUT_FIELD_DEFINITION
-
-directive @federation__override(from: String!) on FIELD_DEFINITION
-
-directive @federation__provides(fields: federation__FieldSet!) on FIELD_DEFINITION
-
-directive @federation__requires(fields: federation__FieldSet!) on FIELD_DEFINITION
-
-directive @federation__shareable on OBJECT | FIELD_DEFINITION
-
-directive @federation__tag(name: String!) repeatable on SCALAR | OBJECT | FIELD_DEFINITION | ARGUMENT_DEFINITION | INTERFACE | UNION | ENUM | ENUM_VALUE | INPUT_OBJECT | INPUT_FIELD_DEFINITION
-
-directive @key(fields: federation__FieldSet!, resolvable: Boolean = true) repeatable on OBJECT | INTERFACE
-
-directive @link(as: String, import: [federation__Import], url: String!) repeatable on SCHEMA
-
-interface Product @key(fields : "id", resolvable : true) {
-  id: ID!
-  name: String!
-}
-
-union _Entity = Book | Movie
-
-type Book implements Product {
-  author: String!
-  id: ID!
-  name: String!
-}
-
-type Movie implements Product {
-  director: String!
-  id: ID!
-  name: String!
-}
-
-type Query {
-  _entities(representations: [_Any!]!): [_Entity]!
-  _service: _Service!
-  movie(id: ID!): Movie
-}
-
-type _Service {
-  sdl: String!
-}
-
-scalar _Any
-
-scalar federation__FieldSet
-
-scalar federation__Import
